Patent/source code/GUI/requirements.txt
pandas==1.3.3
numpy==1.17.3
nltk==3.6.2
tk-tools==0.14.0
tkintertable==1.3.3
beautifulsoup4==4.10.0
tqdm==4.62.2
opencv-python==4.2.0.34
pytesseract==0.3.8
python-pptx==0.6.21
imutils==0.5.4
pdf2image==1.16.0